Add a premake flag to control whether glslang is built from source
This change adds the flag, but doesn't wire it up to anything.
The idea is to allow us to modify CI build scripts to use the flag in cases where it will be needed, so that we can later on disable building glslang from source by default, without an intermediate step where CI is broken.