Fix a bug in handling explicit register space bindings
The logic for handling explicit `space`/`set` bindings on shader parameters for parameter blocks was not correctly marking the `space`/`set` that gets grabbed as used, and as a result it was possible for another parameter block that relies on implicit assignment to end up with a conflicting space.

This change fixes the original oversight, and adds a test case to prevent against regression.